+++ src/encoded/static/components/dbxref.js
@@ -65,9 +65,9 @@ export const dbxrefPrefixMap = {
     HGNC: {
         pattern: 'http://www.genecards.org/cgi-bin/carddisp.pl?gene={0}',
         preprocessor: (context) => {
-            // For dbxrefs in targets, use target.gene_name instead of the dbxref value.
-            if (context['@type'][0] === 'Target' && context.gene_name) {
-                return { altValue: context.gene_name };
+            // For dbxrefs in targets, use the symbol of the first gene in target.genes instead of the dbxref value.
+            if (context['@type'][0] === 'Target' && context.genes && context.genes.length > 0) {
+                return { altValue: context.genes[0].symbol };
             }
             // If a gene displays its dbxrefs, use HGNC URL as NCBI Entrez does.
             if (context['@type'][0] === 'Gene') {
@@ -118,7 +118,7 @@ export const dbxrefPrefixMap = {
         },
     },
     BDSC: {
-        pattern: 'http://flystocks.bio.indiana.edu/Reports/{0}',
+        pattern: 'http://flystocks.bio.indiana.edu/stocks/{0}',
     },
     WormBase: {
         pattern: 'http://www.wormbase.org/species/c_elegans/gene/{0}',