+++ src/encoded/static/components/objectutils.js
@@ -139,7 +139,7 @@ export function requestObjects(atIds, uri, filteringObjects) {
     // complete.
     return Promise.all(objectChunks.map((objectChunk) => {
         // Build URL containing file search for specific files for each chunk of files.
-        const url = uri.concat(objectChunk.reduce((combined, current) => `${combined}&@id=${current}`, ''));
+        const url = uri.concat(objectChunk.reduce((combined, current) => `${combined}&${globals.encodedURIComponent('@id')}=${globals.encodedURIComponent(current)}`, ''));
         return fetch(url, {
             method: 'GET',
             headers: {
@@ -528,7 +528,7 @@ export class BrowserSelector extends React.Component {
 
         return (
             <div className="browser-selector__actuator">
-                <button onClick={this.openModal} disabled={disabled} className="btn btn-info btn-sm" >{title ? <span>{title}</span> : <span>Visualize</span>}</button>
+                <button onClick={this.openModal} disabled={disabled} className="btn btn-info btn-sm" data-test="visualize">{title ? <span>{title}</span> : <span>Visualize</span>}</button>
                 {this.state.selectorOpen ?
                     <Modal closeModal={this.closeModal} addClasses="browser-selector__modal">
                         <ModalHeader title="Open visualization browser" closeModal={this.closeModal} />