+++ src/encoded/types/file.py
@@ -1,4 +1,5 @@
 from botocore.exceptions import ClientError
+from botocore.config import Config
 from snovault import (
     AfterModified,
     BeforeModified,
@@ -25,7 +26,9 @@ from urllib.parse import (
     parse_qs,
     urlparse,
 )
+import base64
 import boto3
+import botocore
 import datetime
 import logging
 import json
@@ -39,6 +42,12 @@ def show_upload_credentials(request=None, context=None, status=None):
     if request is None or status not in ('uploading', 'upload failed'):
         return False
     return request.has_permission('edit', context)
+
+
+def show_cloud_metadata(status=None, md5sum=None, file_size=None, restricted=None):
+    if restricted or not md5sum or not file_size:
+        return False
+    return status in File.public_s3_statuses
 
 
 def property_closure(request, propname, root_uuid):
@@ -331,6 +340,55 @@ class File(Item):
     })
     def superseded_by(self, request, superseded_by):
         return paths_filtered_by_status(request, superseded_by)
+
+    @calculated_property(
+        condition=show_cloud_metadata,
+        schema={
+            "title": "Cloud metadata",
+            "description": "Metadata required for cloud transfer.",
+            "comment": "Do not submit. Values are calculated from file metadata.",
+            "type": "object",
+            "notSubmittable": True,
+        }
+    )
+    def cloud_metadata(self, md5sum, file_size):
+        try:
+            external = self._get_external_sheet()
+        except HTTPNotFound:
+            return None
+        conn = boto3.client('s3', config=Config(
+            signature_version=botocore.UNSIGNED,
+        ))
+        location = conn.generate_presigned_url(
+            ClientMethod='get_object',
+            Params={
+                'Bucket': external['bucket'],
+                'Key': external['key']
+            },
+            ExpiresIn=0
+        )
+        return {
+            'url': location,
+            'md5sum_base64': base64.b64encode(bytes.fromhex(md5sum)).decode("utf-8"),
+            'file_size': file_size
+        }
+
+    @calculated_property(
+        condition=show_cloud_metadata,
+        schema={
+            "title": "S3 URI",
+            "description": "The S3 URI of public file object.",
+            "comment": "Do not submit. Value is calculated from file metadata.",
+            "type": "string",
+            "notSubmittable": True,
+        }
+    )
+    def s3_uri(self):
+        try:
+            external = self._get_external_sheet()
+        except HTTPNotFound:
+            return None
+        return 's3://{bucket}/{key}'.format(**external)
 
     @classmethod
     def create(cls, registry, uuid, properties, sheets=None):