+++ src/encoded/types/target.py
@@ -8,7 +8,9 @@ from .base import (
 )
 from pyramid.traversal import (
     find_root,
+    resource_path
 )
+from snovault.validation import ValidationFailure
 
 
 @collection(
@@ -21,12 +23,40 @@ from pyramid.traversal import (
 class Target(SharedItem):
     item_type = 'target'
     schema = load_schema('encoded:schemas/target.json')
-    embedded = ['organism']
+    embedded = ['organism', 'genes']
 
     def unique_keys(self, properties):
         keys = super(Target, self).unique_keys(properties)
         keys.setdefault('target:name', []).append(self._name(properties))
         return keys
+
+    @calculated_property(schema={
+        "title": "Organism",
+        "description": "Organism bearing the target.",
+        "comment": "Calculated from either target_organism or genes",
+        "type": "string",
+        "linkTo": "Organism"
+    })
+    def organism(self, properties=None, return_uuid=False):
+        if properties is None:
+            properties = self.upgrade_properties()
+        root = find_root(self)
+        if 'target_organism' in properties:
+            organism_uuid = properties['target_organism']
+        else:
+            organism_uuids = set(
+                root.get_by_uuid(gene).upgrade_properties()['organism']
+                for gene in properties['genes']
+            )
+            if len(organism_uuids) != 1:
+                msg = 'Target genes are from different organisms: {}'.format(
+                    repr(organism_uuids)
+                )
+                raise ValidationFailure('body', ['genes'], msg)
+            organism_uuid = next(iter(organism_uuids))
+        if return_uuid:
+            return organism_uuid
+        return resource_path(root.get_by_uuid(organism_uuid), '')
 
     @calculated_property(schema={
         "title": "Name",
@@ -35,11 +65,11 @@ class Target(SharedItem):
     def name(self):
         return self.__name__
 
-    @calculated_property(schema={
+    @calculated_property(condition='organism', schema={
         "title": "Title",
         "type": "string",
     })
-    def title(self, request, organism, label):
+    def title(self, request, label, organism):
         organism_props = request.embed(organism, '@@object')
         return u'{} ({})'.format(label, organism_props['scientific_name'])
 
@@ -50,7 +80,7 @@ class Target(SharedItem):
 
     def _name(self, properties):
         root = find_root(self)
-        organism = root.get_by_uuid(properties['organism'])
+        organism = root.get_by_uuid(self.organism(properties=properties, return_uuid=True))
         organism_props = organism.upgrade_properties()
         return u'{}-{}'.format(properties['label'], organism_props['name'])
 
@@ -58,6 +88,5 @@ class Target(SharedItem):
         request._linked_uuids.add(str(self.uuid))
         # Record organism uuid in linked_uuids so linking objects record
         # the rename dependency.
-        properties = self.upgrade_properties()
-        request._linked_uuids.add(str(properties['organism']))
+        request._linked_uuids.add(str(self.organism(return_uuid=True)))
         return None