+++ src/encoded/upgrade/target.py
@@ -1,3 +1,4 @@
+from pyramid.traversal import find_root
 from snovault import upgrade_step
 
 
@@ -157,3 +158,22 @@ def target_7_8(value, system):
     # Shouldn't be any of these.
     elif status == 'replaced':
         value['status'] = 'deleted'
+
+
+@upgrade_step('target', '8', '9')
+def target_8_9(value, system):
+    # https://encodedcc.atlassian.net/browse/ENCD-3998
+    value.pop('gene_name', '')
+    gene_id_str = 'GeneID:'
+    gene_collection = 'genes'
+    root = find_root(system['context'])
+    genes = [
+        str(root.get(gene_collection).get(dbxref.replace(gene_id_str, '', 1)).uuid)
+        for dbxref in value.get('dbxref', [])
+        if dbxref.startswith(gene_id_str)
+    ]
+    if genes:
+        value['genes'] = genes
+        value.pop('organism')
+    else:
+        value['target_organism'] = value.pop('organism')